+++ igrfCoeffs.h
@@ -1,7 +1,7 @@
 #ifndef _IGRF_COEFFS_H
 #define _IGRF_COEFFS_H
 
-extern const double igrf_coeffs[],igrf_sv[];
+extern const float igrf_coeffs[],igrf_sv[];
 extern const int igrf_date;
 
 extern const int igrf_ord,sv_ord;
